+++ packages/blog/gatsby-blog-core/src/templates/collection.category.js
@@ -18,14 +18,13 @@ export const pageQuery = graphql`
       slug
       description
     }
-
     posts: allArticle(
       filter: {
         private: { ne: true }
         draft: { ne: true }
         category: { slug: { eq: $slug } }
       }
-      sort: { fields: [date], order: DESC }
+      sort: { date: DESC }
       limit: $limit
       skip: $skip
     ) {