+++ packages/common/gatsby-common-helpers/src/default.social.icons.js
@@ -6,25 +6,13 @@ import {
   FaFacebook,
   FaEnvelope,
   FaLink,
-  FaGlobe,
-  FaYoutube,
-  FaCodeBranch
+  FaGlobe
 } from 'react-icons/fa'
-import {BsChatSquareText} from "react-icons/bs";
 
 export default {
   default: {
     Icon: FaLink,
     color: '#718096'
-  },
-  youtube: {
-    Icon: FaYoutube,
-  },
-  matrix: {
-    Icon: BsChatSquareText,
-  },
-  codeberg: {
-    Icon: FaCodeBranch
   },
   github: {
     Icon: FaGithub,
@@ -32,6 +20,7 @@ export default {
   },
   twitter: {
     Icon: FaTwitter,
+    color: '#1DA1F2'
   },
   instagram: {
     Icon: FaInstagram,