+++ packages/common/gatsby-common-helpers/src/useForm.jsx
@@ -22,7 +22,6 @@ const useForm = () => {
 
     // Mimicking form submission for demos
     if (isDemo) {
-      console.info(target, "Demo")
       setTimeout(() => {
         setSuccess(true)
         setSubmitting(false)