+++ packages/flow-ui/flow-ui-components/src/Mdx/CodeBlock.jsx
@@ -1,12 +1,13 @@
-import React from 'react'
-import loadable from '@loadable/component'
+import React, { Suspense } from 'react'
 import { Styled } from 'theme-ui'
 
-const Prism = loadable(() => import('@theme-ui/prism'))
+export const Prism = React.lazy(() => import('@theme-ui/prism'))
 
 const CodeBlock = props => (
   // eslint-disable-next-line react/jsx-pascal-case
-  <Prism {...props} fallback={<Styled.pre>{props.children}</Styled.pre>} />
+  <Suspense fallback={<Styled.pre>{props.children}</Styled.pre>}>
+    <Prism {...props} />
+  </Suspense>
 )
 
 export default CodeBlock