+++ packages/flow-ui/flow-ui-components/src/PoweredByGatsby/index.js
@@ -1,4 +1,1 @@
-import loadable from '@loadable/component'
-export default loadable(() =>
-  import(/* webpackPrefetch: true */ './PoweredByGatsby')
-)
+export { default } from './PoweredByGatsby'