+++ packages/flow-ui/flow-ui-widgets/src/Post/Post.Tags.Share.jsx
@@ -1,6 +1,8 @@
-import React from 'react'
+import React, { Suspense } from 'react'
 import { Flex, Box, Divider } from 'theme-ui'
-import { PostTags, PostShare } from '@widgets/Post'
+import { PostTags } from '@widgets/Post/Post.Tags'
+
+const PostShare = React.lazy(() => import('@widgets/Post/Post.Share'))
 
 const styles = {
   flex: {
@@ -20,7 +22,9 @@ export const PostTagsShare = props => (
     <Divider />
     <Flex sx={styles.flex}>
       <PostTags {...props} />
-      <PostShare {...props} />
+      <Suspense fallback={null}>
+        <PostShare {...props} />
+      </Suspense>
     </Flex>
   </Box>
 )