+++ packages/flow-ui/flow-ui-widgets/src/Search/index.jsx
@@ -1,14 +1,14 @@
-import React, { useState, useCallback } from 'react'
-import loadable from '@loadable/component'
+import React, { useState, useCallback, useTransition, Suspense } from 'react'
 import SearchInput from './Search.Input'
 
-const SearchComponent = loadable(() => import('./Search'))
+const SearchComponent = React.lazy(() => import('./Search'))
 
 const Search = () => {
+  const [, startTransition] = useTransition()
   const [searchLoaded, setSearchLoaded] = useState()
 
   const loadSearchModule = useCallback(() => {
-    SearchComponent.load().then(() => {
+    startTransition(() => {
       setSearchLoaded(true)
     })
   })
@@ -18,7 +18,9 @@ const Search = () => {
   })
 
   return searchLoaded ? (
-    <SearchComponent isFocused={true} />
+    <Suspense fallback={null}>
+      <SearchComponent isFocused={true} />
+    </Suspense>
   ) : (
     <SearchInput loadSearch={loadSearch} />
   )