+++ packages/flow-ui/flow-ui-widgets/src/Seo/Seo.jsx
@@ -31,7 +31,9 @@ const Seo = ({
   const imageSrc = getSrc(getImageVariant(thumbnail, 'hero'))
   const imageUrl =
     imageSrc &&
-    (imageSrc.startsWith('//') ? imageSrc : siteUrl && `${siteUrl}${imageSrc}`)
+    (imageSrc.startsWith('//')
+      ? 'https:' + imageSrc
+      : siteUrl && `${siteUrl}${imageSrc}`)
 
   /**
    * Meta Tags
@@ -48,11 +50,12 @@ const Seo = ({
     { property: 'og:site_name', content: site.name },
     { property: 'og:image', content: imageUrl },
 
-    { name: 'twitter:card', content: 'summary' },
+    { name: 'twitter:card', content: 'summary_large_image' },
     { name: 'twitter:site', content: site.name },
     { name: 'twitter:title', content: title },
     { name: 'twitter:description', content: description },
-    { name: 'twitter:creator', content: twitter.url }
+    { name: 'twitter:creator', content: twitter.url },
+    { name: 'twitter:image', content: imageUrl }
   ]
 
   if (keywords && keywords.length > 0) {