+++ packages/flow-ui/flow-ui-widgets/src/Stats/Stats.jsx
@@ -8,7 +8,7 @@ const styles = {
     bg: `omegaLighter`,
     position: `relative`,
     zIndex: 2,
-    p: 3,
+    p: 3
   },
   icon: {
     display: `flex`,
@@ -17,20 +17,20 @@ const styles = {
     mx: 3,
     svg: {
       color: `omegaDark`,
-      fontSize: 7,
-    },
+      fontSize: 7
+    }
   },
   number: {
     textAlign: `center`,
     fontWeight: `bold`,
     color: `alphaDark`,
     mx: 3,
-    mb: 0,
+    mb: 0
   },
   text: {
     color: `omegaDark`,
-    mb: 0,
-  },
+    mb: 0
+  }
 }
 const Stats = () => {
   return (
@@ -40,19 +40,19 @@ const Stats = () => {
       </Heading>
       <Heading variant='h2' sx={styles.number}>
         250+
-        <Heading variant='h4' sx={styles.text}>
+        <Heading as='p' variant='h4' sx={styles.text}>
           LESSONS
         </Heading>
       </Heading>
       <Heading variant='h2' sx={styles.number}>
         30+
-        <Heading variant='h4' sx={styles.text}>
+        <Heading as='p' variant='h4' sx={styles.text}>
           COURSES
         </Heading>
       </Heading>
       <Heading variant='h2' sx={styles.number}>
         15+
-        <Heading variant='h4' sx={styles.text}>
+        <Heading as='p' variant='h4' sx={styles.text}>
           TUTORS
         </Heading>
       </Heading>