+++ packages/flow-ui/flow-ui-components/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.0](https://github.com/ElegantStack/gatsby-themes/compare/@elegantstack/flow-ui-components@3.4.0...@elegantstack/flow-ui-components@4.0.0) (2023-03-11)
+
+
+### Features
+
+* upgrade to Gatsby v5 and update other dependencies ([aba9eb2](https://github.com/ElegantStack/gatsby-themes/commit/aba9eb243ebe39f0f02885b9693703a316c0a91f))
+
+
+### Bug Fixes
+
+* add React Suspense to fix hydration error ([95a16bc](https://github.com/ElegantStack/gatsby-themes/commit/95a16bcfde3839a957006cd57c91c46229deda06))
+
+
+### Refactor
+
+* refactor code for Gatsby v5 support ([f3852fd](https://github.com/ElegantStack/gatsby-themes/commit/f3852fd3b880cb7b77271f358b092ce730797334))
+
+
+
+
 ## [3.4.0](https://github.com/ElegantStack/gatsby-themes/compare/@elegantstack/flow-ui-components@3.3.0...@elegantstack/flow-ui-components@3.4.0) (2022-05-18)